Restructure all stats functionalities into a separate module
