+++ asgardpy/data/__init__.py
@@ -24,14 +24,10 @@ from asgardpy.data.target import (
     ExpCutoffLogParabolaSpectralModel,
     Target,
     apply_selection_mask_to_models,
-    check_model_preference_aic,
-    check_model_preference_lrt,
     config_to_dict,
     create_gal_diffuse_skymodel,
     create_iso_diffuse_skymodel,
     create_source_skymodel,
-    get_chi2_sig_pval,
-    get_goodness_of_fit_stats,
     set_models,
     xml_spatial_model_to_gammapy,
     xml_spectral_model_to_gammapy_params,
@@ -61,14 +57,10 @@ __all__ = [
     "FluxPointsConfig",
     "Target",
     "apply_selection_mask_to_models",
-    "check_model_preference_aic",
-    "check_model_preference_lrt",
     "config_to_dict",
     "create_gal_diffuse_skymodel",
     "create_iso_diffuse_skymodel",
     "create_source_skymodel",
-    "get_chi2_sig_pval",
-    "get_goodness_of_fit_stats",
     "set_models",
     "xml_spatial_model_to_gammapy",
     "xml_spectral_model_to_gammapy_params",
