+++ asgardpy/stats/__init__.py
@@ -1,0 +1,18 @@
+"""
+Statistics Module
+"""
+from asgardpy.stats.stats import (
+    check_model_preference_aic,
+    check_model_preference_lrt,
+    get_chi2_sig_pval,
+    get_goodness_of_fit_stats,
+    get_ts_null_hypothesis,
+)
+
+__all__ = [
+    "check_model_preference_aic",
+    "check_model_preference_lrt",
+    "get_chi2_sig_pval",
+    "get_goodness_of_fit_stats",
+    "get_ts_null_hypothesis",
+]
