+++ docs/source/_api_docs/data_target_f.rst
@@ -2,7 +2,7 @@ asgardpy.data.target module: Functions
 ======================================
 
 .. automodule:: asgardpy.data.target
-   :members: set_models, config_to_dict, create_source_skymodel, params_renaming_to_gammapy, params_rescale_to_gammapy, xml_spectral_model_to_gammapy_params, xml_spatial_model_to_gammapy, create_gal_diffuse_skymodel, create_iso_diffuse_skymodel, apply_selection_mask_to_models, check_model_preference_lrt, check_model_preference_aic, get_chi2_sig_pval, get_goodness_of_fit_stats
+   :members: set_models, config_to_dict, create_source_skymodel, params_renaming_to_gammapy, params_rescale_to_gammapy, xml_spectral_model_to_gammapy_params, xml_spatial_model_to_gammapy, create_gal_diffuse_skymodel, create_iso_diffuse_skymodel, apply_selection_mask_to_models
    :undoc-members:
    :show-inheritance:
    :noindex:
