+++ docs/source/_api_docs/stats.rst
@@ -1,0 +1,7 @@
+asgardpy.stats.stats module
+===========================
+
+.. automodule:: asgardpy.stats.stats
+   :members: check_model_preference_lrt, check_model_preference_aic, get_chi2_sig_pval, get_goodness_of_fit_stats, get_ts_null_hypothesis
+   :undoc-members:
+   :show-inheritance:
